fix windows case
